requirements.txt
textual>=0.80.0
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
html2text>=2024.2.26
rapidfuzz>=3.6.0
keyring>=24.3.0
